# polymorphic-api/src/main/java/me/anantagarwal/dto/PaymentResponse.java
package me.anantagarwal.dto;

import com.fasterxml.jackson.annotation.JsonSubTypes;
import com.fasterxml.jackson.annotation.JsonTypeInfo;

@JsonTypeInfo(
        use = JsonTypeInfo.Id.NAME,
        include = JsonTypeInfo.As.PROPERTY,
        property = "type"
)
@JsonSubTypes({
        @JsonSubTypes.Type(value = CreditCardPaymentResponse.class, name = "credit_card"),
        @JsonSubTypes.Type(value = BankTransferPaymentResponse.class, name = "bank_transfer")
})
public abstract class PaymentResponse {
    private String id;
    private double amount;
    private String currency;
    private String status;

    // Getters and setters
    public String getId() {
        return id;
    }

    public void setId(String id) {
        this.id = id;
    }

    public double getAmount() {
        return amount;
    }

    public void setAmount(double amount) {
        this.amount = amount;
    }

    public String getCurrency() {
        return currency;
    }

    public void setCurrency(String currency) {
        this.currency = currency;
    }

    public String getStatus() {
        return status;
    }

    public void setStatus(String status) {
        this.status = status;
    }
}